update github workflows

+++ ios_bitcode.sh
@@ -6,12 +6,15 @@ NEW_WRAP=$2
 cd ~
 echo "=====[ Getting Depot Tools ]====="	
 git clone -q https://chromium.googlesource.com/chromium/tools/depot_tools.git
-cd depot_tools
-git reset --hard 8d16d4a
-cd ..
+if [ "$VERSION" != "10.6.194" -a "$VERSION" != "11.8.172" ]; then 
+    cd depot_tools
+    git reset --hard 8d16d4a
+    cd ..
+fi
 export DEPOT_TOOLS_UPDATE=0
 export PATH=$(pwd)/depot_tools:$PATH
 gclient
+~/depot_tools/ensure_bootstrap
 
 
 mkdir v8
@@ -23,6 +26,7 @@ echo "target_os = ['ios']" >> .gclient
 cd ~/v8/v8
 git checkout refs/tags/$VERSION
 gclient sync
+python3 tools/clang/scripts/update.py
 
 
 echo "=====[ Patching V8 ]====="
